requirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.7
django-celery-beat==2.2.0
django-chartjs==2.2.1
django-crispy-forms==1.11.2
django-crontab==0.7.1
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.12.3
djangorestframework-simplejwt==5.0.0
greenlet==1.1.0
gunicorn==20.0.4
idna==3.2
importlib-metadata==3.10.0
kombu==5.1.0
Markdown==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2==2.8.6
PyJWT==2.3.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
reportlab==3.5.68
requests==2.26.0
scipy==1.5.4
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.6
vcversioner==2.16.0.0
vine==5.0.0
wcwidth==0.2.5
xlwt==1.3.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
